refactor(archive-metadata): move the metadata writer out of the src root
`patchArchiveMetadata` composes the ComicInfo and OPF container modules
the same way those modules compose their own read/write halves, so it
belongs beside them rather than at the package root.
