+++ packages/archive-metadata/src/index.ts
@@ -13,7 +13,7 @@ export type {
   ArchiveMetadataTargets,
   ArchivePatch,
   ArchivePatchedEntry,
-} from "./writer"
-export { patchArchiveMetadata } from "./writer"
+} from "./metadata/write"
+export { patchArchiveMetadata } from "./metadata/write"
 
 export { normalizeIsbn } from "@prose-reader/archive-reader"
